+++ ledger/sandbox-on-x/BUILD.bazel
@@ -46,7 +46,6 @@ da_scala_library(
         "//ledger/participant-state",
         "//ledger/participant-state-index",
         "//ledger/participant-state-metrics",
-        "//ledger/sandbox-common",
         "//libs-scala/build-info",
         "//libs-scala/concurrent",
         "//libs-scala/contextualized-logging",
@@ -207,7 +206,6 @@ da_scala_test_suite(
         "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-integration-api:participant-integration-api-tests-lib",
-        "//ledger/sandbox-common",
         "//ledger/test-common",
         "//libs-scala/contextualized-logging",
         "//libs-scala/grpc-utils",
