+++ ledger/sandbox-on-x/src/main/scala/com/daml/ledger/sandbox/Banner.scala
@@ -1,13 +1,13 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.sandbox.banner
+package com.daml.ledger.sandbox
 
 import java.io.PrintStream
-
 import scala.io.Source
 
 object Banner {
+
   private val classLoader = getClass.getClassLoader
 
   private val resourceName = "banner.txt"
@@ -19,4 +19,5 @@ object Banner {
       .mkString("\n")
 
   def show(out: PrintStream): Unit = out.println(banner)
+
 }
